+++ src/main/java/fr/isika/projet3/service/UserService.java
@@ -1,13 +1,14 @@
 package fr.isika.projet3.service;
 
 import java.util.List;
+
 import fr.isika.projet3.entities.User;
 
 public interface UserService  {
 
 	public List getAllUsers();
-	public User getUserById(Long id);
+	public User getUserById(int id);
 	public boolean saveUser(User user);
-	public boolean deleteUserById(Long id);
+	public boolean deleteUserById(int id);
 
 }
